requirements.txt
Django==3.1
mock==4.0.2
ipython==7.17.0
smart-open==2.1.0
boto3==1.14.48
moto==1.3.14
pre-commit==2.7.1
django-model-utils==4.0.0
celery==4.4.7
requests==2.24.0
factory-boy==3.0.1
exam==0.10.6
responses==0.11.0
django-dirtyfields==1.4